Split stage module into package
