+++ tests/integration/test_project_navigator.py
@@ -3,6 +3,7 @@ Tests for project navigator and its views
 """
 import logging
 import textwrap
+
 import ddt
 import mock
 
@@ -13,7 +14,7 @@ from group_project_v2.stage import (
     BasicStage, SubmissionStage, TeamEvaluationStage, PeerReviewStage,
     EvaluationDisplayStage, GradeDisplayStage, CompletionStage
 )
-from group_project_v2.utils import StageState
+from group_project_v2.stage.utils import StageState
 from tests.integration.base_test import SingleScenarioTestSuite, BaseIntegrationTest
 from tests.integration.page_elements import NavigationViewElement, ResourcesViewElement, SubmissionsViewElement, \
     GroupProjectElement
